Update moray-firth.md
Radioskills Summer School - This project is being made available for up to 12 participants, over 16 years of age and to anyone with an interest or aspiration to learn new digital skills. Radioskills Summer School candidates will learn how to create entertaining digital content that will connect and engage with audiences both at home and abroad. They will learn how to write and plan a digital radio programme as well as how to use industry standard digital recording and editing equipment. 
It is likely that with the participants permission we would be able to showcase their broadcast on our Inverness based trainees radio station, Monster FM,  in September.